feat(arguments): automatically pass through arguments to yarn
Deprecate "--cmd" and "--opt" options, and automatically pass through arguments to yarn when they
are not already handled by yarn-recursive. For example, the command `yarn-recursive --cmd run
--option build` is now simplified to just `yarn-recursive run build`.

BREAKING CHANGE: The "--cmd" and "--opt" options are deprecated. They are still supported, but their
behaviour is subtly different to that of previous versions. Users are encouraged to switch to the
new syntax as soon as possible.